pyproject.toml
[tool.poetry]
name = "you-owe-me"
version = "0.1.0"
description = "Tinkin Technical debt analyzer"
authors = ["Tinkin"]
readme="README.md"
packages = [
    { include = "you-owe-me" },
    { include = "you-owe-me/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"

[tool.poetry.scripts]
dev = "src:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
